requirements.txt
click==6.6
cookiecutter==1.6.0
ruamel.yaml==0.16.10
pygithub==1.51
GitPython==3.1.2
autopep8==1.5.2
cffi==1.14.0
cryptography==2.9.2
requests==2.23.0
rich==1.1.9
packaging==20.4
